KFC - 97: account filtering by department
Added a drop down menu that allows a user to select the department they want to filter by. Had to create a separate test array of accounts as currently our accounts do not contain a department parameter. However, testing worked completely with the dummy set of accounts. Functionality won't work on the main branch until we get departments linked to accounts.